[REFACTOR] Split up graphics definitions
